created reschedule job

+++ db/db_actions.py
@@ -86,9 +86,10 @@ async def get_tasks_from_db() -> list[Remind]:
 # удалить задание из бд
 async def delete_task_from_db(job: apscheduler.events.JobEvent):
     session: AsyncSession = db_helper.get_scoped_session()
-    result = await session.execute(select(Remind).where(Remind.id == int(job.job_id)))
-    if task := result.scalar():
-        logger.info(f"удален job с id: {job.job_id}")
-        await session.delete(task)
-    await session.commit()
-    await session.close()
+    if job.job_id.isdigit():
+        result = await session.execute(select(Remind).where(Remind.id == int(job.job_id)))
+        if task := result.scalar():
+            logger.info(f"удален job с id: {job.job_id}")
+            await session.delete(task)
+        await session.commit()
+        await session.close()
